+++ index.html
@@ -7,7 +7,12 @@
     <script>
       yawl.configure('cda712a73aff22114b6f62871697ea15')
 
-      yawl.track("click");
+      yawl.track("click", {
+        article_id: 69,
+        establishment_account_id: 109,
+        name: 'test',
+        user_type: 'client'
+      });
     </script>
   </head>
   <body>
